% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxandwhiskers.R
\name{boxandwhiskers}
\alias{boxandwhiskers}
\title{Box And Whiskers}
\usage{
boxandwhiskers(data, d, tutorialMode)
}
\arguments{
\item{data}{Input data.}

\item{d}{Degree of outlier or distance at which an event is considered an outlier}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\description{
This function implements the box & whiskers algorithm to detect outliers
}
\examples{
boxandwhiskers(inputData, 2, FALSE)
boxandwhiskers(c(1,2,3,4,5,6,7), 1, TRUE)

}
