% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{knn}
\usage{
knn(data, d, K, tutorialMode)
}
\arguments{
\item{data}{Input Data (must be a data.frame)}

\item{d}{Degree of outlier or distance at which an event is considered an outlier}

\item{K}{Nearest neighbor for which an event must have a degree of outlier to be considered an outlier}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\description{
This function implements the knn algorithm for outlier detection
}
\author{
Andres Missiego Manjon
}
