% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.R
\name{lof}
\alias{lof}
\title{lof}
\usage{
lof(inputData, K, d, tutorialMode)
}
\arguments{
\item{inputData}{Input Data (must be a data.frame)}

\item{K}{This number represents the nearest neighbor to use to calculate the density of each point. This value is chosen arbitrarily and is responsibility of the data scientist/user to select a number adequate to the dataset.}

\item{d}{degree of outlier, this value is multiplied by the standard deviation and added/subtracted to the mean to obtain the limit values as it has been explained earlier in this section. This value is chosen arbitrarily and is responsibility of the data scientist/user to select a number adequate to the dataset.}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\description{
Local Outlier Factor algorithm to detect outliers
}
\author{
Andres Missiego Manjon
}
