% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_distance.R
\name{mahalanobis_distance}
\alias{mahalanobis_distance}
\title{mahalanobis_distance}
\usage{
mahalanobis_distance(value, sample_mean, sample_covariance_matrix)
}
\arguments{
\item{sample_mean}{Sample mean}

\item{sample_covariance_matrix}{Sample Covariance Matrix}

\item{point}{Point to calculate the mahalanobis_distance}
}
\value{
Mahalanobis distance associated to the point
}
\description{
Calculates the mahalanobis_distance given the input data
}
