% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_method.R
\name{mahalanobis_method}
\alias{mahalanobis_method}
\title{mahalanobis_distance}
\usage{
mahalanobis_method(inputData, alpha, tutorialMode)
}
\arguments{
\item{inputData}{Input Data dataset}

\item{alpha}{Significance level alpha}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\description{
Detect outliers using the Mahalanobis Distance
}
\examples{
inputData = t(matrix(c(3,2,3.5,12,4.7,4.1,5.2,4.9,7.1,6.1,6.2,5.2,14,5.3),2,7,dimnames=list(c("r","d"))));
inputData = data.frame(inputData);
mahalanobis_method(inputData, 0.7, FALSE);

inputData = t(matrix(c(3,2,3.5,12,4.7,4.1,5.2,4.9,7.1,6.1,6.2,5.2,14,5.3),2,7,dimnames=list(c("r","d"))));
inputData = data.frame(inputData);
mahalanobis_method(inputData, 0.2, TRUE);

}
\author{
Andres Missiego Manjon
}
