% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_dist.R
\name{manhattan_dist}
\alias{manhattan_dist}
\title{manhattan_dist}
\usage{
manhattan_dist(A, B)
}
\arguments{
\item{A}{One of the 2D points}

\item{B}{The other 2D point}
}
\value{
Manhattan distance calculated between point A and B
}
\description{
Calculates the manhattan distance between two 2D points
}
\examples{
inputData = t(matrix(c(3,2,3.5,12,4.7,4.1,5.2,4.9,7.1,6.1,6.2,5.2,14,5.3),2,7,dimnames=list(c("r","d"))));
inputData = as.matrix(inputData);
distance = manhattan_dist(inputData[3,], inputData[4,]);
print(distance);

distance = manhattan_dist(c(1,2), c(3,4));
print(distance);

}
\author{
Andres Missiego Manjon
}
