% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_outliersLearn.R
\name{mean_outliersLearn}
\alias{mean_outliersLearn}
\title{mean_outliersLearn}
\usage{
mean_outliersLearn(data)
}
\arguments{
\item{data}{Input Data that will be processed to calculate the mean. It must be a vector}
}
\value{
Mean of the input data
}
\description{
Calculates the mean of the given data vector
}
\examples{
inputData = c(1,2,3,4,5,6,1);
mean = mean_outliersLearn(inputData);
print(mean);

mean = mean_outliersLearn(c(2,3,2.3,7.8));
print(mean);

}
\author{
Andres Missiego Manjon
}
