% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_method.R
\name{sd_method}
\alias{sd_method}
\title{sd_method}
\usage{
sd_method(data, d, tutorialMode)
}
\arguments{
\item{data}{Input Data}

\item{d}{Degree of outlier or distance at which an event is considered an outlier}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\description{
This function implements the outlier detection algorithm using standard deviation and mean
}
\examples{
inputData = t(matrix(c(3,2,3.5,12,4.7,4.1,5.2,4.9,7.1,6.1,6.2,5.2,14,5.3),2,7,dimnames=list(c("r","d"))))
inputData = data.frame(inputData)
sd_method(inputData,2,TRUE)

sd_method(c(1,2,3,4,1,23,4),2,FALSE)

}
\author{
Andres Missiego Manjon
}
