% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_outliersLearn.R
\name{sd_outliersLearn}
\alias{sd_outliersLearn}
\title{sd_outliersLearn}
\usage{
sd_outliersLearn(data, mean)
}
\arguments{
\item{data}{Input Data that will be used to calculate the standard deviation. Must be a vector}

\item{mean}{Mean of the input data vector of the function.}
}
\value{
Standard Deviation of the input data
}
\description{
Calculates the standard deviation of the input data given the mean.
}
\examples{
inputData = c(1,2,3,4,5,6,1);
mean = sum(inputData)/length(inputData);
sd = sd_outliersLearn(inputData, mean);

}
\author{
Andres Missiego Manjon
}
