% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxandwhiskers.R
\name{boxandwhiskers}
\alias{boxandwhiskers}
\title{Box And Whiskers}
\usage{
boxandwhiskers(data, d, tutorialMode)
}
\arguments{
\item{data}{Input data.}

\item{d}{Degree of outlier or distance at which an event is considered an outlier}

\item{tutorialMode}{if TRUE the tutorial mode is activated (the algorithm will include an explanation detailing the theory behind the outlier detection algorithm and a step by step explanation of how is the data processed to obtain the outliers following the theory mentioned earlier)}
}
\value{
None, does not return any value
}
\description{
This function implements the box & whiskers algorithm to detect outliers
}
\examples{
inputData = t(matrix(c(3,2,3.5,12,4.7,4.1,5.2,
4.9,7.1,6.1,6.2,5.2,14,5.3),2,7,dimnames=list(c("r","d"))))
inputData = data.frame(inputData)
boxandwhiskers(inputData,2,FALSE) # Can be set to TRUE

}
\author{
Andres Missiego Manjon
}
