% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_dist.R
\name{manhattan_dist}
\alias{manhattan_dist}
\title{manhattan_dist}
\usage{
manhattan_dist(A, B)
}
\arguments{
\item{A}{One of the 2D points}

\item{B}{The other 2D point}
}
\value{
Manhattan distance calculated between point A and B
}
\description{
Calculates the manhattan distance between two 2D points
}
\examples{
distance = manhattan_dist(c(1,2), c(3,4));

}
\author{
Andres Missiego Manjon
}
