% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_outliersLearn.R
\name{mean_outliersLearn}
\alias{mean_outliersLearn}
\title{mean_outliersLearn}
\usage{
mean_outliersLearn(data)
}
\arguments{
\item{data}{Input Data that will be processed to calculate the mean. It must be a vector}
}
\value{
Mean of the input data
}
\description{
Calculates the mean of the given data vector
}
\examples{
mean = mean_outliersLearn(c(2,3,2.3,7.8));

}
\author{
Andres Missiego Manjon
}
