% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_outliersLearn.R
\name{quantile_outliersLearn}
\alias{quantile_outliersLearn}
\title{quantile_outliersLearn}
\usage{
quantile_outliersLearn(data, v)
}
\arguments{
\item{data}{Input Data}

\item{v}{Goes from 0 to 1 (e.g. 0.25). Indicates the quantile that wants to be obtained}
}
\value{
Quantile v calculated
}
\description{
Function that obtains the 'v' quantile
}
\examples{
q = quantile_outliersLearn(c(12,2,3,4,1,13), 0.60)

}
\author{
Andres Missiego Manjon
}
